Hide validator-js from browserify and webpack.
When a simple string is provided to `require` WebPack and browserify will attempt to find it to add it to the bundle.
Since the `npm` name is different it will not be found in the local `node_modules`.
By making this an expression browserify and WebPack will ignore this require.